# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WeedTown API",
    "version": "2.4.0",
    "description": "API de WeedTown, red social de la comunidad cannábica de México. Identidad federada con Mastodon (OAuth 2.0) y sesión mediante JWT propio en el header Authorization. Seguridad: headers via helmet, CORS restringido al frontend, rate limiting (300 peticiones/15 min general; 20/15 min en el flujo OAuth), payload JSON limitado a 100 kB e imágenes a 5 MB. El perfil público (GET /api/profile/{id}) no expone PII (email, teléfono, nombre real, edad, nacimiento ni género). Las rutas de mercado y admin son stubs pendientes de implementación."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT emitido por el backend tras el flujo OAuth de Mastodon (expira en 7 días)."
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "mastodonInstance": {
            "type": "string",
            "example": "mastodon.social"
          },
          "acct": {
            "type": "string",
            "example": "usuario@mastodon.social"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "avatar": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "age": {
            "type": "integer",
            "nullable": true
          },
          "birthdate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true,
            "enum": [
              "masculino",
              "femenino",
              "otro",
              null
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/api/auth/mastodon/start": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Inicia el flujo OAuth con una instancia de Mastodon",
        "description": "Registra dinámicamente la app en la instancia si es la primera vez y redirige a su pantalla de autorización con un state firmado (expira en 10 minutos).",
        "parameters": [
          {
            "name": "instance",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "mastodon.social"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirección a https://{instance}/oauth/authorize"
          },
          "400": {
            "description": "Instancia inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/mastodon/callback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Callback OAuth (uso interno del flujo)",
        "description": "Canjea el código por un access token, obtiene el perfil con verify_credentials, hace upsert del usuario y redirige al frontend con el JWT en el fragmento: {FRONTEND_URL}/auth/callback#token=...",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirección al frontend con el token, o a /login?error=... si falla"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Usuario de la sesión actual",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Sin token o token inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Incluye `role` y, si aplica, `suspendedUntil` y `suspendedReason`. El rol viaja para que el frontend sepa si pintar el panel; la autorización real se verifica en el servidor en cada petición."
      }
    },
...
    "/api/blocks": {
      "get": {
        "tags": [
          "Seguridad"
        ],
        "summary": "Cuentas que bloqueé",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Solo los bloqueos propios. Quien fue bloqueado no puede consultarlos ni enterarse de que existen.",
        "responses": {
          "200": {
            "description": "Lista de cuentas bloqueadas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "blocks": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "name": {
                                "type": "string"
                              },
                              "displayName": {
                                "type": "string",
                                "nullable": true
                              },
                              "avatar": {
                                "type": "string",
                                "nullable": true
                              },
                              "acct": {
                                "type": "string"
                              }
                            }
                          },
                          {
                            "type": "object",
                            "properties": {
                              "blockedAt": {
                                "type": "string",
                                "format": "date-time"
                              }
                            }
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          }
        }
      },
      "post": {
        "tags": [
          "Seguridad"
        ],
        "summary": "Bloquear a una persona (idempotente)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Efecto mutuo mientras dure: ninguna de las dos partes ve ni puede contactar a la otra en feed, foros, chat y Cerca. Al bloqueado no se le avisa. Además borra las notificaciones que esa persona había generado.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bloqueo activo"
          },
          "400": {
            "description": "userId faltante o intento de bloquearte a ti"
          },
          "401": {
            "description": "No autenticado"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      }
    },
    "/api/blocks/{userId}": {
      "delete": {
        "tags": [
          "Seguridad"
        ],
        "summary": "Desbloquear (idempotente)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bloqueo retirado"
          },
          "400": {
            "description": "userId inválido"
          },
          "401": {
            "description": "No autenticado"
          }
        }
      }
    },
...
  }
}